move: enchant or potion matter interfaces moved to their respective packages
