+++ api/src/main/java/io/github/sakaki_aruka/customcrafter/matter/enchant/CEnchantBothMatterImpl.kt
@@ -1,7 +1,7 @@
 package io.github.sakaki_aruka.customcrafter.matter.enchant
 
-import io.github.sakaki_aruka.customcrafter.matter.CEnchantMatter
-import io.github.sakaki_aruka.customcrafter.matter.CEnchantmentStoreMatter
+import io.github.sakaki_aruka.customcrafter.matter.enchant.CEnchantMatter
+import io.github.sakaki_aruka.customcrafter.matter.enchant.CEnchantmentStoreMatter
 import io.github.sakaki_aruka.customcrafter.matter.CMatter
 import io.github.sakaki_aruka.customcrafter.matter.CMatterPredicate
 import io.github.sakaki_aruka.customcrafter.matter.enchant.CEnchantComponent
