+++ api/src/main/java/io/github/sakaki_aruka/customcrafter/matter/potion/CPotionMatterImpl.kt
@@ -2,7 +2,7 @@ package io.github.sakaki_aruka.customcrafter.matter.potion
 
 import io.github.sakaki_aruka.customcrafter.matter.CMatter
 import io.github.sakaki_aruka.customcrafter.matter.CMatterPredicate
-import io.github.sakaki_aruka.customcrafter.matter.CPotionMatter
+import io.github.sakaki_aruka.customcrafter.matter.potion.CPotionMatter
 import io.github.sakaki_aruka.customcrafter.matter.potion.CPotionComponent
 import io.github.sakaki_aruka.customcrafter.matter.CMatterImpl
 import org.bukkit.Material
